oauth: factor out sessionIssuerHelper
