--- enterprise/cmd/frontend/auth/githuboauth/user.go
@@ -1,1 +1,0 @@
-package githuboauth
