+++ enterprise/cmd/frontend/auth/gitlaboauth/session.go
@@ -1,0 +1,73 @@
+package gitlaboauth
+
+import (
+	"context"
+	"fmt"
+	"net/http"
+	"strconv"
+
+	"github.com/pkg/errors"
+	"github.com/sourcegraph/sourcegraph/cmd/frontend/db"
+	"github.com/sourcegraph/sourcegraph/cmd/frontend/external/auth"
+	"github.com/sourcegraph/sourcegraph/enterprise/cmd/frontend/auth/oauth"
+	"github.com/sourcegraph/sourcegraph/pkg/actor"
+	"github.com/sourcegraph/sourcegraph/pkg/extsvc"
+	"github.com/sourcegraph/sourcegraph/pkg/extsvc/gitlab"
+	"golang.org/x/oauth2"
+)
+
+type sessionIssuerHelper struct {
+	*gitlab.CodeHost
+	clientID string
+}
+
+func (s *sessionIssuerHelper) GetOrCreateUser(ctx context.Context, token *oauth2.Token) (actr *actor.Actor, safeErrMsg string, err error) {
+	gUser, err := UserFromContext(ctx)
+	if err != nil {
+		return nil, "Could not read GitLab user from callback request.", errors.Wrap(err, "could not read user from context")
+	}
+
+	login, err := auth.NormalizeUsername(gUser.Username)
+	if err != nil {
+		return nil, fmt.Sprintf("Error normalizing the username %q. See https://docs.sourcegraph.com/admin/auth/#username-normalization.", login), err
+	}
+
+	var data extsvc.ExternalAccountData
+	data.SetAccountData(gUser)
+	data.SetAuthData(token)
+
+	userID, safeErrMsg, err := auth.CreateOrUpdateUser(ctx, db.NewUser{
+		Username:        login,
+		Email:           gUser.Email,
+		EmailIsVerified: gUser.Email != "",
+		DisplayName:     gUser.Name,
+		AvatarURL:       gUser.AvatarURL,
+	}, extsvc.ExternalAccountSpec{
+		ServiceType: s.ServiceType(),
+		ServiceID:   s.ServiceID(),
+		ClientID:    s.clientID,
+		AccountID:   strconv.FormatInt(int64(gUser.ID), 10),
+	}, data)
+	if err != nil {
+		return nil, safeErrMsg, err
+	}
+	return actor.FromUser(userID), "", nil
+}
+
+func (s *sessionIssuerHelper) DeleteStateCookie(w http.ResponseWriter) {
+	stateConfig := getStateConfig()
+	stateConfig.MaxAge = -1
+	http.SetCookie(w, oauth.NewCookie(stateConfig, ""))
+}
+
+func (s *sessionIssuerHelper) SessionData(token *oauth2.Token) oauth.SessionData {
+	return oauth.SessionData{
+		ID: auth.ProviderConfigID{
+			ID:   s.ServiceID(),
+			Type: s.ServiceType(),
+		},
+		AccessToken: token.AccessToken,
+		TokenType:   token.Type(),
+		// TODO(beyang): store and use refresh token to auto-refresh sessions
+	}
+}
