+++ schema/schema.go
@@ -173,7 +173,8 @@ type GitHubAuthProvider struct {
 
 // GitHubAuthorization description: If non-null, enforces GitHub repository permissions. This requires that there is an item in the `auth.providers` field of type "github" with the same `url` field as specified in this `GitHubConnection`.
 type GitHubAuthorization struct {
-	Ttl string `json:"ttl,omitempty"`
+	Organizations []string `json:"organizations,omitempty"`
+	Ttl           string   `json:"ttl,omitempty"`
 }
 type GitHubConnection struct {
 	Authorization               *GitHubAuthorization `json:"authorization,omitempty"`
